requirements.txt
langchain==0.1.1
streamlit==1.30.0
streamlit-chat==0.1.1
openai==1.10.0
langchain-openai==0.0.5
# requests
beautifulsoup4 == 4.12.2
tiktoken
facebook-sdk